requirements.txt
simple-sds011==0.0.7
pyserial>=3.5
tomli>=2.0.1; python_version < "3.11"
prometheus_client>=0.19.0
